feat: add plant registration form
Made plantFormInputValidation to validate the input fields of the plant registration form.
Made registerPlant to send the data to the backend.
Made PlantRegister to render the form and handle the submit event.
Made PlantContainer to handle the submit event and send the data to the backend and display the form component